requirements.txt
bottle
python-magic
